requirements.txt
openai
flask
python-dotenv
flask-cors
whisper
httpx
certify
openai
